Add account data to appointments

+++ rfh_server/apps/appointments/models.py
@@ -20,7 +20,7 @@ class AppointmentsDB(models.Model):
     appointmentState = models.CharField(max_length=250, default='non')
     summary = models.CharField(max_length=550, default='non')
     mpesaPaymentId = models.CharField(max_length=550, default='non')
-    amountPayed = models.IntegerField(default=0)
+    amountPayed = models.FloatField(default=0.0)
     appointmentType = models.IntegerField(default=0)
     timestamp = models.DateTimeField(default=datetime.now)
